Made FFT window size customizable at runtime

+++ include/utils.h
@@ -7,7 +7,7 @@ namespace SamplesToNotes
 
 FftwReal FrequencyOfNote(size_t note, FftwReal a4Frequency = 440);
 FftwReal DifferencePercentage(FftwReal value1, FftwReal value2);
-void GenerateSinWave(FftwReal* buffer, int bufferSize, FftwReal frequency, FftwReal samplingRate);
+void GenerateSinWave(FftwReal* buffer, size_t bufferSize, FftwReal frequency, FftwReal samplingRate);
 
 static const char* NoteNames[12] =
 {
